Using python library function for gpio access
